MUS-22: fix some codestyle issues, add dummy auth repo interface
also:
 - rename some directories from snake_case to dash-case,
   some files, that are including those directories as well.
 - add dummy auth repo interface for supression compilation errors
   of AuthUseCase class.
 - fix indentation for namespaces due to codestyle guide.
