+++ lib/mus-internal/mus-irepo/include/mus-irepo/ichat_repository.h
@@ -1,7 +1,9 @@
 #ifndef MUS_INTERNAL_MUS_IREPO_ICHAT_REPOSITORY_H_
 #define MUS_INTERNAL_MUS_IREPO_ICHAT_REPOSITORY_H_
 
-#include "mus_models/chat.h"
+#include <vector>
+
+#include "mus-models/chat.h"
 #include "irepository.h"
 
 namespace music_share {
