+++ lib/mus-internal/mus-models/include/mus-models/chat_message.h
